backend/requirement.txt
fastapi
uvicorn
openai
cohere
supabase
python-dotenv
tiktoken
